+++ src/transport/encoding/raptorq/encoder.rs
@@ -13,12 +13,14 @@ use super::super::Configurable;
 
 const DEFAULT_MIN_REPAIR_PACKETS_PER_BLOCK: u32 = 5;
 const DEFAULT_MTU: u16 = 1400;
+const DEFAULT_FEQ_REDUNDANCY: f32 = 0.15;
 
 use raptorq::Encoder as ExtEncoder;
 
 pub(crate) struct RaptorQEncoder {
     min_repair_packets_per_block: u32,
     mtu: u16,
+    fec_redundancy: f32,
 }
 
 impl Configurable for RaptorQEncoder {
@@ -34,9 +36,17 @@ impl Configurable for RaptorQEncoder {
             .map(|s| s.parse().ok())
             .flatten()
             .unwrap_or(DEFAULT_MTU);
+
+        let fec_redundancy = conf
+            .get("fec_redundancy")
+            .map(|s| s.parse().ok())
+            .flatten()
+            .unwrap_or(DEFAULT_FEQ_REDUNDANCY);
+
         Self {
             min_repair_packets_per_block,
             mtu,
+            fec_redundancy,
         }
     }
     fn default_configuration() -> HashMap<String, String> {
@@ -61,9 +71,9 @@ impl Encoder for RaptorQEncoder {
             let mut base_packet = payload.generate_uid().to_vec();
             base_packet.append(&mut transmission_info);
 
-            let mut repair_packets = (payload.gossip_frame.len() * 10
-                / (self.mtu as usize)
-                / 100) as u32;
+            let mut repair_packets = (payload.gossip_frame.len() as f32
+                * self.fec_redundancy
+                / self.mtu as f32) as u32;
             if repair_packets < self.min_repair_packets_per_block {
                 repair_packets = self.min_repair_packets_per_block
             }